changed lower upper bound validation to parseFloat instead of parseInt

+++ public/js/controllers/newFactoryController.js
@@ -46,7 +46,7 @@ function newFactoryController($scope, $mdDialog) {
 
   $scope.validateLowerBound = function name() {
     console.log($scope.newFactoryData.lowerBound, typeof $scope.newFactoryData.lowerBound);
-    if (typeof parseInt($scope.newFactoryData.lowerBound) === 'number' && $scope.newFactoryData.lowerBound !== undefined && !isNaN(parseInt($scope.newFactoryData.lowerBound)) && isFinite(parseInt($scope.newFactoryData.lowerBound))) {
+    if (typeof parseFloat($scope.newFactoryData.lowerBound) === 'number' && $scope.newFactoryData.lowerBound !== undefined && !isNaN(parseFloat($scope.newFactoryData.lowerBound)) && isFinite(parseFloat($scope.newFactoryData.lowerBound))) {
       $scope.lowerBoundInValid = false;
       console.log("lowerBound is valid");
       return true;
@@ -59,7 +59,7 @@ function newFactoryController($scope, $mdDialog) {
 
   $scope.validateUpperBound = function name() {
     console.log($scope.newFactoryData.upperBound, typeof $scope.newFactoryData.upperBound);
-    if (typeof parseInt($scope.newFactoryData.upperBound) === 'number' && $scope.newFactoryData.upperBound !== undefined && !isNaN(parseInt($scope.newFactoryData.upperBound)) && isFinite(parseInt($scope.newFactoryData.upperBound))) {
+    if (typeof parseFloat($scope.newFactoryData.upperBound) === 'number' && $scope.newFactoryData.upperBound !== undefined && !isNaN(parseFloat($scope.newFactoryData.upperBound)) && isFinite(parseFloat($scope.newFactoryData.upperBound))) {
       $scope.upperBoundInValid = false;
       console.log("upperBound is valid");
       return true;
